hand1 update
I’ve added a base hand that links up! What is the industry standard for
rigging hands? I always lock the finger controls to the rig arm and
affect the main joints directly (that way the hands for IK and FK are
the same controls). Is that weird or acceptable?  I ask ‘cause I’ve
been rigging that way for a while and I found it faster than setting up
IK/FK digits by hand.