updated for week 3
Added nested loop to create joint chains.  Changed IK control to a square.  Having issues getting the pivot of the control to center on the joint.  Added the pole vector.  Tried adding the IK switch as a text, but having problems getting that to work correctly.  I've commented out the areas in question and added more of an explanation.  And I still can't get GitHub Desktop to do what I want.