classes aren't writing to json file
not entirely sure how to get this so iI can track down where things are
going wrong. It builds 90% correctly, but my values never make it to my
JSON file, so i can’t check what indices I need to get it to work
correctly. it won’t parent the digits or their controls to the hand any
more.