it would seem that I broke something worse.
I think part of the reason it never gets to Json is because killer bugs
show up as I fix the ones that don’t seem to be as bad. like this new
one that claims invalid syntax in kc5 on line 188 all the other for
loops are set up the same, but this one stops everything from running.
The pycharm errors that appear are talking about “End of statement
expected” and I’m not sure what it means by that. in maya where it
claims syntax error, there’s a teeny carrot under the ‘:’ but i’m not
sure what that means. Aaaaaahhhh I’m breaking things and I don’t
understand how or whyyyyyyy :P Also, I did try to rearrange the file
structure, but then it lost the file locations altogether and didn’t
print out any errors. I will dig into that more, but I have the more
concerns about the sudden death of kc5.